+++ package.json
@@ -16,7 +16,7 @@
     "start": "yarn workspace @botpress/messaging-server start",
     "dev": "yarn workspace @botpress/messaging-server dev",
     "board": "yarn workspace @botpress/messaging-board dev",
-    "test": "yarn workspace @botpress/messaging-server test"
+    "test": "jest"
   },
   "bin": "./packages/server/dist/index.js",
   "pkg": {
@@ -32,6 +32,7 @@
     "outputPath": "./bin"
   },
   "devDependencies": {
+    "@types/jest": "^27.0.1",
     "@typescript-eslint/eslint-plugin": "^4.29.1",
     "@typescript-eslint/parser": "^4.29.1",
     "conventional-changelog": "^3.1.24",
@@ -41,11 +42,13 @@
     "eslint-plugin-import": "^2.24.0",
     "eslint-plugin-jsdoc": "^35.5.1",
     "fs-extra": "^10.0.0",
+    "jest": "^27.0.6",
     "parcel": "^2.0.0-beta.3.1",
     "pkg": "^4.3.7",
     "prettier": "^2.3.2",
     "rimraf": "^3.0.2",
     "semver": "^7.3.5",
+    "ts-jest": "^27.0.4",
     "ts-node-dev": "^1.1.8",
     "typescript": "^4.3.5",
     "yargs": "^17.0.1"
