+++ packages/server/package.json
@@ -7,7 +7,6 @@
     "build": "yarn --silent && tsc --build",
     "watch": "yarn --silent && tsc --build --watch",
     "start": "node dist/index.js",
-    "test": "jest",
     "dev": "yarn --silent && ts-node-dev --inspect --inspect-publish-uid=http --transpile-only src/index.ts"
   },
   "devDependencies": {
@@ -15,15 +14,12 @@
     "@types/cli-color": "^2.0.1",
     "@types/express": "^4.17.13",
     "@types/ioredis": "^4.26.7",
-    "@types/jest": "^27.0.1",
     "@types/lodash": "^4.14.172",
     "@types/lru-cache": "^5.1.1",
     "@types/ms": "^0.7.31",
     "@types/node": "^12.18.3",
     "@types/redlock": "^4.0.2",
-    "@types/uuid": "^8.3.1",
-    "jest": "^27.0.6",
-    "ts-jest": "^27.0.4"
+    "@types/uuid": "^8.3.1"
   },
   "dependencies": {
     "@botpress/messaging-base": "0.0.1",
